// files/src/react/modules/bnpl-plan-config/bnpl.enum.ts
export enum PlanStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
}

export enum PlanType {
  MONTHLY = 'monthly',
  ANNUALLY = 'annual',
}

export enum Currency {
  MYR = 'MYR',
}

export enum Country {
  MALAYSIA = 'MY',
}

export enum PlanStructure {
  INSTRUCTION = 'instruction',
  POSTPAID = 'postpaid',
}